Cấu hình token cho notification-service

+++ BACKEND/notification-service/src/main/java/fpt/aptech/notificationservice/dtos/NotifyDTO.java
@@ -5,8 +5,9 @@ import lombok.*;
 @Data
 public class NotifyDTO {
     private int itemId;
-    private Long userId;
+    private int userId;
     private String fullName;
     private String title;
     private String content;
+   // private String token;
 }
